bifrost/bifrost/requirements.txt
Django==4.2.7
django-filter==23.4
djangorestframework==3.14.0
Markdown==3.5.1
sqlparse==0.4.4
typing_extensions==4.8.0
psycopg2-binary==2.9.6
drf-yasg==1.21.7